refactor SampleQueries into own component
